♻️  router reporting.py 추출 — work report + clarification display (P0-P step 10)
변경 이유
- ``_emit_work_report_preview`` + ``_format_clarification_message`` +
  outcome / report coercion 4 함수가 ``_legacy`` 안에서 책임이 다른
  주변 코드와 entangled 되어 있었다. 출력 / 직렬화 책임이라 하나로
  묶어 단독 모듈에 분리.

주요 변경 사항
- 새 파일 ``reporting.py`` (약 200 줄):
  - ``_emit_work_report_preview`` — WorkReport 빌드 + 저장 + Discord 게시
  - ``_format_clarification_message`` — ACTION_ASK 분기 후보 디스플레이
  - ``_coerce_outcome`` / ``_coerce_research_loop_report`` — 방어적
    shape 변환
- ``_legacy.py`` 의 해당 4 함수 (195 줄) 제거 → re-export.
- ``research_loop.py`` 의 ``_coerce_research_loop_report`` lazy shim
  유지 (reporting 이 caller, research_loop 이 callee 라는 의존이
  깨지지 않게).
- ``_legacy.py`` 최상단에 utils / research_loop / reporting 의 local
  binding 재import 블록 추가 — _legacy 내부 함수들의 call site 가
  여전히 local name 으로 호출하므로 모두 가용해야 함.
- ``__init__.py`` facade 가 ``.reporting`` 에서 ``_coerce_outcome`` /
  ``_emit_work_report_preview`` 직접 import.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- ``_legacy.py`` 1583 → 1399 줄 (P0-P 시작 3316 줄에서 누적 1917 줄 감소).
